proposta_form_completo_v2/requirements.txt
Flask==3.0.3
python-docx==1.1.2
Werkzeug==3.0.3
